requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
matplotlib>=3.7.0
Pillow>=10.0.0
natsort>=8.4.0
pycocotools>=2.0.7
numpy>=1.24.0
